WARNING!!! This machine is subject to breakdowns during periods of critical need.
A special circuit in the machine called "critical detector" senses the
operator's emotional state in terms of how desperate he/she is to use the
machine.  The "critical detector" then creates a malfunction proportional
to the desperation of the operator.  Threatening the machine with violence
only aggravates the situation.  Likewise, attempts to use another machine
may cause it to malfunction.  They belong to the same union.  Keep cool
and say nice things to the machine.  Nothing else seems to work.

See also: flog(1), tm(1)